requirements.txt
# Quality Evaluation Dependencies
pyyaml>=6.0
strands-agents
nest-asyncio
tenacity>=8.0.0
boto3>=1.26.0
rebrowser-playwright>=1.40.0
